EC-105-Modified some return types and added a method to delete an order item
Changed some return types to match the controller, and deleted the method that retrieves the order items because it turned out that I could access the order items for each order using the orderItems list assigned to each order.
I also changed the library from which we use the JsonIgnore annotation because the old one was not breaking the cyclic dependency between the order and the order item, and I moved the annotation from the order's orderItems to orderItem's order (if that made any sense) because I need the orderItems to be serialized, but I don't need the order for the order item itself (Only one is enough, and the list of order items is more useful for me).